+++ Assets/UniSharper.Data.Metadata/CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [3.0.0] - 2026-04-29
+
+### Added
+
+- Adds interface **IDatabaseCryptoProvider**.
+- Adds interface **IDatabaseCompressionProvider**.
+- Supports database encryption and compression.
+- Supports custom database crypto provider and compression provider.
+
+
+
 ## [2.6.0] - 2026-04-02
 
 ### Changed